KP-8461: Render list of questions based on type

+++ src/UI/Headquarters/WB.UI.Headquarters.Interview/src/components/questions/Index.ts
@@ -1,11 +1,11 @@
 import * as vue from "vue"
 
 import DateTime from "./DateTime"
-import Numeric from "./Numeric"
+import Integer from "./Integer"
 import SingleOption from "./SingleOption"
 import TextQuestion from "./TextQuestion"
 
 vue.component("DateTime", DateTime)
-vue.component("Numeric", Numeric)
+vue.component("Integer", Integer)
 vue.component("TextQuestion", TextQuestion)
 vue.component("SingleOption", SingleOption)
